requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==25.1.0
autopep8==2.3.1
backports.tarfile==1.2.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==44.0.0
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==4.4.2
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docopt==0.6.2
docutils==0.21.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.0.1
ffmpeg==1.4
FLAML==2.1.2
Flask==3.0.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.0.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==25.2.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.41
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langsmith==0.3.11
Mako==1.3.9
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.3.0
moviepy==1.0.3
multidict==6.1.0
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nh3==0.2.18
numpy==1.26.4
oauthlib==3.2.2
openai==1.61.1
openpyxl==3.1.5
orjson==3.10.15
packaging==24.1
pafy==0.5.5
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.10.0
platformdirs==4.2.2
proglog==0.1.10
prompt_toolkit==3.0.47
propcache==0.3.0
proto-plus==1.24.0
protobuf==5.27.2
psutil==6.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyautogen==0.2.33
pycodestyle==2.12.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic-settings==2.8.1
pydantic_core==2.20.1
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-vlc==3.0.20123
pytube==15.0.0
pytubefix==6.5.1
pytz==2025.1
PyYAML==6.0.1
pyzmq==26.0.3
readme_renderer==44.0
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rsa==4.9
SecretStorage==3.3.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
tenacity==9.0.0
termcolor==2.4.0
tiktoken==0.7.0
tomli==2.2.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
twine==5.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
yarl==1.18.3
youtube-dl==2021.12.17
youtube-transcript-api==0.6.2
zipp==3.19.2
zstandard==0.23.0
